Changed format of timer from a simple message to an embed, and changed it to where users can now only use one timer at a time for performance purposes.

+++ modules/clockModule.py
@@ -14,17 +14,20 @@ commandList = []
 commandList.append(Command("!stopwatch", "stopwatch", "TODO"))
 async def stopwatch(client, message):
     pass
-# start = time.time()
-# while int(time.time() - start) < int(10):
-#     time.sleep(1)
-#     print(math.floor(time.time() - start))
-# Helper to get the most recent timer
-async def getLastTimer(client, message):
-    hist = await message.channel.history(limit=30).flatten()
+    # start = time.time()
+    # while int(time.time() - start) < int(10):
+    #     time.sleep(1)
+    #     print(math.floor(time.time() - start))
+    # Helper to get the most recent timer
+async def getLastTimer(message):
+    hist = await message.channel.history(limit=100).flatten()
+    lastTimer = None
     for old in hist:
-        if old.author == client.user and (old.content.find("remaining.") != -1):
-            lastTimer = old
-            break
+        if old.embeds:
+            description = old.embeds[0].description
+            if description.find("remaining.") != -1:
+                lastTimer = old
+                break
     return lastTimer
 
 commandList.append(Command("!timer", "timer", "Used to start, pause, unpause, or delete a timer.\nTo create a new timer, use: `!timer <TIME>`. The amount of time given is in minutes, (ex: if you want a 30 second timer use 0.5)\nTo pause the most recent timer, use: `!timer pause`.\nTo unpause the most recent timer, use: `!timer unpause`.\nTo delete the most recent timer, use: `!timer delete`."))
@@ -34,19 +37,32 @@ async def timer(client, message):
         await message.channel.send("Please specify the amount of time you want on your timer and whether it is in minutes or seconds.\nUsage: !timer <min/sec> <TIME>")
         return
     if content[1] == 'delete':
-        lastTimer = await getLastTimer(client, message)
+        lastTimer = await getLastTimer(message)
+        if lastTimer == None:
+            await message.channel.send("No timer found.")
+            return
         await lastTimer.delete()
         await message.channel.send("Timer successfully deleted.")
         return
     elif content[1] == 'pause':
-        lastTimer = await getLastTimer(client, message)
-        timeLeft = lastTimer.content.split(" ")[0]
+        lastTimer = await getLastTimer(message)
+        if lastTimer == None:
+            await message.channel.send("No timer found.")
+            return
+        description = lastTimer.embeds[0].description
+        timeLeft = description.split(" ")[0]
         await lastTimer.delete()
-        await message.channel.send(content="{} remaining.".format(timeLeft))
+        msg = "{} remaining.".format(timeLeft)
+        embed = discord.Embed(title = "Timer", description=msg, colour = discord.Colour.purple())
+        await message.channel.send(embed=embed)
         return
     elif content[1] == 'unpause':
-        lastTimer = await getLastTimer(client, message)
-        left = lastTimer.content.split(" ")[0]
+        lastTimer = await getLastTimer(message)
+        if lastTimer == None:
+            await message.channel.send("No timer found.")
+            return
+        description = lastTimer.embeds[0].description
+        left = description.split(" ")[0]
         start = time.time()
         end = float(left.split(":")[0]) + (float(left.split(":")[1]) / 60)
         while int(time.time() - start) < math.floor(float(end) * 60):
@@ -55,19 +71,19 @@ async def timer(client, message):
                 left = str(math.floor(secondsLeft / 60)) + ":0" + str(secondsLeft % 60)
             else:
                 left = str(math.floor(secondsLeft / 60)) + ":" + str(secondsLeft % 60)
+            msg = "{} remaining.".format(left)
+            embed = discord.Embed(title = "Timer", description=msg, colour = discord.Colour.purple())
             try:
-                await lastTimer.edit(content="{} remaining.".format(left))
+                await lastTimer.edit(embed=embed)
             except:
                 return
             time.sleep(0.95)
-        await lastTimer.edit(content="0:00 remaining.")
+        msg = "0:00 remaining."
+        embed = discord.Embed(title = "Timer", description=msg, colour = discord.Colour.purple())
+        await lastTimer.edit(embed=embed)
         await message.channel.send(message.author.mention + ", your timer is complete.")
         return
-        
-        
-    # if not(content[1] == "min" or content[1] == "sec"):
-    #     await message.channel.send("Please specify the amount of time you want on your timer and whether it is in minutes or seconds.\nUsage: !timer <min/sec> <TIME>")
-    #     return
+
     end = content[1]
     def is_number(s):
         if s.isnumeric():
@@ -80,31 +96,31 @@ async def timer(client, message):
     if not is_number(end):
         await message.channel.send("Please enter an amount of time to run a timer.")
         return
+    # Delete old timer
+    old = await getLastTimer(message)
+    if old:
+        await old.delete()
     start = time.time()
-    # if content[1] == 'sec':
-    #     newMessage = await message.channel.send("{} seconds remaining.".format(end))
-    #     while int(time.time() - start) < math.floor(float(end)):
-    #         await newMessage.edit(content="{} seconds remaining.".format(int(math.floor(float(end))) - math.floor(time.time() - start)))
-    #         time.sleep(0.95)
-    #     try:
-    #         await newMessage.edit(content="0 seconds remaining.")
-    #     except:
-    #         return
-    #elif content[1] == 'min':
     left = "0:00"
-    newMessage = await message.channel.send("Starting your timer.")
+    msg = "Starting your timer."
+    embed = discord.Embed(title = "Timer", description=msg, colour = discord.Colour.purple())
+    newMessage = await message.channel.send(embed=embed)
     while int(time.time() - start) < math.floor(float(end) * 60):
         secondsLeft = math.floor(float(end) * 60) - math.floor(time.time() - start)
         if (secondsLeft % 60 < 10):
             left = str(math.floor(secondsLeft / 60)) + ":0" + str(secondsLeft % 60)
         else:
             left = str(math.floor(secondsLeft / 60)) + ":" + str(secondsLeft % 60)
+        msg = "{} remaining.".format(left)
+        embed = discord.Embed(title = "Timer", description=msg, colour = discord.Colour.purple())
         try:
-            await newMessage.edit(content="{} remaining.".format(left))
+            await newMessage.edit(embed=embed)
         except:
             return
         time.sleep(0.95)
-    await newMessage.edit(content="0:00 remaining.")
+    msg = "0:00 remaining."
+    embed = discord.Embed(title = "Timer", description=msg, colour = discord.Colour.purple())
+    await newMessage.edit(embed=embed)
     await message.channel.send(message.author.mention + ", your timer is complete.")
